requirements.txt
fastapi==0.101.1
pydantic==2.2.1
PyJWT==1.7.1
PyJWT==2.8.0
python-decouple==3.8
uvicorn==0.23.2
